template-api changes, ehic-api tags changed, dtos simplified
