+++ libs/clients/ehic-client-v1/project.json
@@ -25,5 +25,5 @@
       }
     }
   },
-  "tags": []
+  "tags": ["lib:client", "scope:client"]
 }
